docs: comment pinned add bar
