+++ src/lib/components/index.ts
@@ -6,3 +6,4 @@ export { default as ClearedList } from './ClearedList.svelte';
 export { default as AdvancedDetails } from './AdvancedDetails.svelte';
 export { default as TagsList } from './TagsList.svelte';
 export { default as PriorityFilter } from './PriorityFilter.svelte';
+export { default as AddBar } from './AddBar.svelte';
